LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-007-expected.txt


PASS Should use the last position option initially
FAIL Should stay at initial position, since it still fits assert_equals: Anchored element should be at the top of anchor expected 450 but got 600
PASS Should use the third position option with enough space left
PASS Should use the first position option with enough space left and below
FAIL Should use the second position option with enough space below (and not enough above) assert_equals: Anchored element should be at the right of anchor expected 135 but got 35
